+++ the_project/special_scripts/items.py
@@ -4,21 +4,41 @@ This deals with all the items the player can equip.
 
 import arcade
 import math
+import logging
 from the_project.constants import *
 
 
-class Hammer(arcade.Sprite):
-    """
-    The player hammer. Used to build and heal buildings.
-    """
+class Item(arcade.Sprite):
+    def __init__(self, texture: str, icon_texture: str, icon_x: int, icon_y: int):
+        """
+        This is the class that will be the building block for the rest of the item entities.
+        __variables are not to be altered outside the class.
 
-    def __init__(self):
+        :param string texture: Path to the Texture of the Item.
+        :param string icon_texture: Path to the Texture of the Icon of the Item
+        :param int icon_x: X-Coordinate of Icon
+        :param int icon_y: Y-Coordinate of Icon
+        """
         super().__init__()
-        self.texture = arcade.load_texture("assets/images/other_sprites/hotbar_items/hotbar_item_hammer.png")
+        self.texture = arcade.load_texture(texture)
+        self.__path_to_texture = texture
+        self.__icon = arcade.Sprite(icon_texture)
+        self.__icon.position = (icon_x, icon_y)
+        self.__path_to_icon = icon_texture
 
-        # This is a backup. Mouse hasn't been moved yet, it will default to the player.
+        # This is a backup. If the mouse hasn't been moved yet, it will default to the player.
         window = arcade.get_window()
         self.prev_mouse_x, self.prev_mouse_y = window.scene[SCENE_NAME_BLUE_PLAYER][0].position
+
+    def __repr__(self):
+        """
+        Runs when the entire class is called (e.g. printed)
+
+        :return: A basic report of the Item
+        :rtype: str
+        """
+        return f"Item. Failed to implement __repr__ for the class, so its defaulting to the item __repr__. " \
+               f"Texture: {self.get_texture()}. Icon Texture: {self.get_icon_texture()}"
 
     def on_click(self, x: float, y: float, button: int, modifiers: int):
         """
@@ -30,20 +50,22 @@ class Hammer(arcade.Sprite):
         :param modifiers: Bitwise 'and' of all modifiers (shift, ctrl, num lock) pressed during this event.
         See :ref:`keyboard_modifiers`.
         """
-        # This will not work if the player is red!
-        window = arcade.get_window()
         if button == 1:
-            x2, y2 = window.scene[SCENE_NAME_BLUE_PLAYER][0].position
-            distance = arcade.get_distance(x1=x, y1=y, x2=x2, y2=y2)
-
-            if distance <= 100:
-                click_list = (arcade.get_sprites_at_point((x, y), window.scene[SCENE_NAME_BLUE_BUILDING]))
-                for building in click_list:
-                    building.change_current_health(200)
+            try:
+                self.left_click(x=x, y=y)
+            except AttributeError:
+                logging.error(f"Item.on_click - In - 'Item'. self.left_click() does not exist, as it has not been "
+                              f"implemented. {self.__repr__()}")
+        elif button == 4:
+            try:
+                self.right_click(x=x, y=y)
+            except AttributeError:
+                logging.error(f"Item.on_click - In - 'Item'. self.right_click() does not exist, as it has not been "
+                              f"implemented. {self.__repr__()}")
 
     def update_position(self, mouse_x=None, mouse_y=None):
         """
-        Updates the hammers position. Run every time the mouse is moved.
+        Updates the Items position. Run every time the mouse is moved.
 
         :param float mouse_x: x position of mouse. Default to None if the mouse havent moved at all.
         :param float mouse_y: y position of mouse. Default to None if the mouse havent moved at all.
@@ -72,33 +94,81 @@ class Hammer(arcade.Sprite):
         angle = math.degrees(angle)
         self.angle = -angle
 
+    def draw(self):
+        super().draw()
+        self.__icon.draw()
 
+    def get_texture(self):
+        """
+        :return: Path to the texture
+        :rtype: str
+        """
+        return self.__path_to_texture
 
+    def get_icon_texture(self):
+        """
+        :return: The path to the icon texture
+        :rtype: str
+        """
+        return self.__path_to_texture
 
-        # diff_x = mouse_x - player_x
-        # diff_y = mouse_y - player_y
-        #
-        # combined_diff = diff_x + diff_y
-        # if combined_diff >= 200:
-        #
-        # # if diff_x > 100:
-        # #     diff_x = 100
-        # # elif diff_x < -100:
-        # #     diff_x = -100
-        # #
-        # # if diff_y > 100:
-        # #     diff_y = 100
-        # # elif diff_y < -100:
-        # #     diff_y = -100
-        #
-        # self.center_x = player_x + diff_x
-        # self.center_y = player_y + diff_y
-        # angle = math.atan2(diff_y, diff_x)
-        #
-        # # Set the enemy to face the player.
-        # angle = math.degrees(angle) - 90
-        # self.angle = angle
+    def get_icon_position(self):
+        """
+        :return: Position of the icon.
+        :rtype: int
+        """
+        return self.__icon.position
 
+    def set_icon_position(self, x: int, y: int):
+        """
+        :param x: X-Coordinate of the Icon
+        :param y: Y-Coordinate of the Icon
+        """
+        self.__icon.position = (x, y)
 
     def kill(self):
         super().kill()
+
+
+class Hammer(Item):
+    def __init__(self, icon_x: int, icon_y: int):
+        """
+        The players hammer. Used to build and heal buildings.
+
+        :param int icon_x: X-Coordinate of Icon
+        :param int icon_y: Y-Coordinate of Icon
+        """
+        super().__init__(texture="assets/images/other_sprites/hotbar_items/hotbar_item_hammer.png",
+                         icon_texture="assets/images/other_sprites/hotbar_icons/hotbar_icon_hammer.png",
+                         icon_x=icon_x,
+                         icon_y=icon_y
+                         )
+
+    def __repr__(self):
+        """
+        Runs when the entire class is called (e.g. printed)
+
+        :return: A basic report of the Item
+        :rtype: str
+        """
+        return f"Item - Hammer. ({self.center_x}, {self.center_y}), " \
+               f"Texture: {self.get_texture()}. Icon Texture: {self.get_icon_texture()}"
+
+    def left_click(self, x: float, y: float):
+        """
+        Runs when the left click is called.
+
+        :param x: X-Coord of click.
+        :param y: Y-Coord of click.
+        """
+        # This will not work if the player is red!
+        window = arcade.get_window()
+        x2, y2 = window.scene[SCENE_NAME_BLUE_PLAYER][0].position
+        distance = arcade.get_distance(x1=x, y1=y, x2=x2, y2=y2)
+
+        if distance <= 100:
+            click_list = (arcade.get_sprites_at_point((x, y), window.scene[SCENE_NAME_BLUE_BUILDING]))
+            for building in click_list:
+                building.change_current_health(200)
+
+